FIX: haskell-wrapper-nix: hardcode reference to wrapper
I was previously calling the wrapper by name, and not the full path.
But meant the user needed to confure the wrapper to be on the PATH.
Since the wrapper is portable from project to project, the script
is now hardcoding the reference to the wrapper.  This should only
simply the user experience.